+++ Sources/BraintreeLocalPayment/BTLocalPaymentClient.swift
@@ -24,7 +24,10 @@ import BraintreeDataCollector
     
     /// Exposed for testing to get the instance of BTAPIClient
     var apiClient: BTAPIClient
-    
+
+    /// Exposed for testing to inject mock data collector
+    var dataCollector: BTDataCollector?
+
     // MARK: - Private Properties
 
     private var request: BTLocalPaymentRequest?
@@ -68,9 +71,6 @@ import BraintreeDataCollector
                 return
             }
 
-            let dataCollector = BTDataCollector(authorization: self.apiClient.authorization.originalValue)
-            request.correlationID = dataCollector.clientMetadataID(nil)
-
             guard let configuration else {
                 self.notifyFailure(with: BTLocalPaymentError.fetchConfigurationFailed, completion: completion)
                 return
@@ -89,7 +89,20 @@ import BraintreeDataCollector
                 return
             }
 
-            self.start(request: request, configuration: configuration)
+            let dataCollector = self.dataCollector ?? BTDataCollector(authorization: self.apiClient.authorization.originalValue)
+            dataCollector.collectDeviceData(riskCorrelationID: nil) { deviceData, error in
+                if let error {
+                    self.notifyFailure(with: error, completion: completion)
+                    return
+                }
+
+                if let deviceData, let data = deviceData.data(using: .utf8) {
+                    let json = BTJSON(data: data)
+                    request.correlationID = json["correlation_id"].asString()
+                }
+
+                self.start(request: request, configuration: configuration)
+            }
         }
     }