+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## unreleased
 
+* BraintreeDataCollector
+  * Update PPRiskMagnes to 5.6.0 with internal completion handler to ensure device data is not returned until Magnes SDK completes data submission
+  * Deprecate `BTDataCollector.clientMetadataID` - use `BTDataCollector(riskCorrelationID:completion:)` instead
+
 * BraintreePayPal
     * Pass `fundingSource` to the app switch url link and to analytics events. 
     * Fix bug where setting an empty `BTPayPalVaultRequest.userAuthenticationEmail` resulted in an error